Diabetic disease model
Summary
In model used 5 alogrithms i.e LogisticRegression,Decision tree,RandomForest Gradient boosting & Support vector Machine.
 Accurcay of Logistic Regression,Gradient Boosting & Support Vector Machine is 0.99 that means it is overfitting.So, we will not consider that model as model is overfitting.
Comparsion of accuracy of Decision tree,Random Forest.As random forest accuracy is 0.98 but consider accuracy of decision tree i. 0.96 because if we consider random forest as per accuracy as it has trained & have as we trained  a student bookish knowledge & only having a bokish knowledge can be harmful.For eg,if a student have only bookish & he byheart everything instead of understanding.So,It can be problematic as undertanding & practical knowledge is not necessary.So, it is not suitable for our model.That's why we will consider Decision tree accuracy & Decision tree alagorithms is giving  beter accuracy i.e 0.96.
Conclusion:Decision tree alogrithms is better for the Diabetic disease prediction models.